extracting frontendWindow to it's own file
